pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-data_centric_env"
version = "0.1.0"
description = "Data Centric Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # Environment-specific dependencies
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m data_centric_env.server.app
server = "data_centric_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["data_centric_env", "data_centric_env.server"]
package-dir = { "data_centric_env" = ".", "data_centric_env.server" = "server" }

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]